+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.0] - 2025-07-15
+
+### Added
+
+- Sequence opener sample
+- Generating and using style overrides
+- Generating and using panel settings overrides
+
 ## [1.0.0] - 2025-02-24
 
 ### Added